% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneWayLRTcrit.R
\name{onewayLRTcrit}
\alias{onewayLRTcrit}
\title{Find test statistic value from a given data and corresponding critical value}
\usage{
onewayLRTcrit(data, alpha)
}
\arguments{
\item{data}{a real data set considering only numerical values}

\item{alpha}{real number between 0 and 1 called significance level}
}
\value{
numeric vector
}
\description{
More detailed description
}
\examples{
k=4;N=c(20,30,20,30);S=c(1,1,2,1)
g=NULL
for(i in 1:k)
{
 g[[i]]=rnorm(N[i],0,sqrt(S[i]))
}
data=g
onewayLRTcrit(data,0.05)

}
