% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneWayLRT.R
\docType{package}
\name{OwoWayLRT}
\alias{OwoWayLRT}
\title{Find test statistic value from a given data and corresponding critical value}
\description{
This package finds critical value along with statistic value of bootstrap based LRT method for testing equility of
treatment effects against ordered alternative in One-Way ANOVA model with heteroscedasticity.
}
\author{
Anjana Mondal \email{mondal.anjana25@gmail.com}
}
